move TryReadChars to shared and cleanup
